First Commit.
There's a lot here, so I'll keep it very brief:

Added JDA for Discord and event listeners.
Created a database using Notepad (this was a design choice as many Discord users are not technical).
Implemented the checking component.
Implemented server scanner.
This is an extremely brief summary of what was added, as there was a lot added.